+++ src/types/address-table.types.ts
@@ -1,0 +1,13 @@
+export interface IndexedAccountType {
+  address: string;
+  balance: string;
+  usdValue?: string;
+  index: number;
+}
+
+export interface IAddressTableData {
+  accounts: IndexedAccountType[];
+  startIndex: number;
+  addressesPerPage: number;
+  isLoading: boolean;
+}